+++ src/Ui5/Contracts/Ui5DialogInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace LaravelUi5\Core\Ui5\Contracts;
+
+interface Ui5DialogInterface extends Ui5ArtifactInterface, SluggableInterface
+{
+
+}